+++ backend/substrapp/views/utils.py
@@ -5,6 +5,9 @@ from wsgiref.util import is_hop_by_hop
 
 import django.http
 from django.conf import settings
+from django.forms import TypedChoiceField
+from django_filters.rest_framework import BaseInFilter
+from django_filters.rest_framework import TypedChoiceFilter
 from rest_framework import status
 from rest_framework.authentication import BasicAuthentication
 from rest_framework.filters import SearchFilter
@@ -251,3 +254,9 @@ class MatchFilter(SearchFilter):
 
     def get_search_fields(self, view, request):
         return getattr(view, "search_fields", self.default_search_fields)
+
+
+class TypedChoiceInFilter(BaseInFilter, TypedChoiceFilter):
+    """Allow typed choice field to be filtered with IN lookup passing comma separated values list"""
+
+    field_class = TypedChoiceField